Commit añadiendo directorio en galería donde se guardan las capturas de la cámara.

+++ app/src/main/java/com/example/luis/series/actividades/Perfil.java
@@ -4,13 +4,19 @@ import android.app.Activity;
 import android.content.DialogInterface;
 import android.content.Intent;
 import android.graphics.Bitmap;
+import android.graphics.BitmapFactory;
+import android.media.MediaScannerConnection;
 import android.net.Uri;
+import android.os.Environment;
+import android.os.StrictMode;
 import android.provider.MediaStore;
 import android.support.v7.app.AlertDialog;
 import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
 import android.util.Log;
 import android.view.View;
+import android.view.animation.Animation;
+import android.view.animation.AnimationUtils;
 import android.widget.Button;
 import android.widget.EditText;
 import android.widget.ImageView;
@@ -34,6 +40,7 @@ import com.google.firebase.storage.StorageReference;
 import com.google.firebase.storage.UploadTask;
 
 import java.io.ByteArrayOutputStream;
+import java.io.File;
 import java.text.SimpleDateFormat;
 import java.util.Date;
 
@@ -41,26 +48,31 @@ public class Perfil extends AppCompatActivity {
 
     ImageView imagenUsuario;
     EditText editTextEmailUsuario;
-    private static final int GALLERY_INTENT = 1;
-    private static final int CAMERA_INTENT = 2;
+    private static final int GALLERY_INTENT=4;
+    private static final int CAMERA_INTENT=5;
+    private static final int CAMARA=0;
+    private static final int GALERIA=1;
+    private static final int CANCELAR=2;
     private boolean camara = false, galeria = false, cambios = false;
-    private ByteArrayOutputStream stream;
     private Uri imagenSeleccionada, enlaceFotoFirebase;
     private StorageReference filePath;
     private StorageReference storageReference;
     private Usuario user;
     Button botonModificar;
+    private  final String CARPETA_RAIZ="Series/";
+    private final String RUTA_IMAGEN=CARPETA_RAIZ+"series";
+    String miPath;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_perfil);
+        StrictMode.VmPolicy.Builder builder = new StrictMode.VmPolicy.Builder();
+        StrictMode.setVmPolicy(builder.build());
         botonModificar = findViewById(R.id.botonModificar);
         storageReference = FirebaseStorage.getInstance().getReference();
         imagenUsuario = findViewById(R.id.fotoPerfil);
         editTextEmailUsuario = findViewById(R.id.editTextEmail);
-        //String emailUsuario= getIntent().getStringExtra(TabActivity.EMAIL_USUARIO);
-        //String fotoUsuario=getIntent().getStringExtra(TabActivity.FOTO_USUARIO);
         user = (Usuario) getIntent().getSerializableExtra("usuario");
         Log.i("avatar", "-> " + user.getAvatar());
         Glide.with(this)
@@ -75,31 +87,43 @@ public class Perfil extends AppCompatActivity {
             public void onClick(View view) {
 
 
-                final CharSequence[] items = {"Cámara", "Galería", "Cancelar"};
+                final CharSequence[] items = {getString(R.string.camara),getString(R.string.galeria),getString(R.string.cancelar)};
                 AlertDialog.Builder builder = new AlertDialog.Builder(Perfil.this);
-                builder.setTitle("Selecciona un foto de perfil");
+                builder.setTitle(R.string.selec_accion);
                 builder.setItems(items, new DialogInterface.OnClickListener() {
                     @Override
                     public void onClick(DialogInterface dialogInterface, int i) {
-                        String opcion = (String) items[i];
+                        int opcion= i;
                         switch (opcion) {
 
-                            case "Cámara":
-                                Log.i("perfil", "Selecciona camara");
+                            case CAMARA:
+                                Log.i("perfil","Selecciona camara");
+                                File fileImagen = new File(Environment.getExternalStorageDirectory(),RUTA_IMAGEN);
+                                boolean creada=fileImagen.exists();
+                                String nombreImagen="";
+                                if(!creada){
+                                    creada=fileImagen.mkdirs();
+                                }
+                                if(creada){
+                                    nombreImagen=(System.currentTimeMillis()/1000)+".jpg";
+                                }
+                                miPath=Environment.getExternalStorageDirectory()+File.separator+RUTA_IMAGEN+File.separator+nombreImagen;
+                                File imagen = new File(miPath);
 
                                 Intent intent = new Intent(MediaStore.ACTION_IMAGE_CAPTURE);
-                                startActivityForResult(intent, CAMERA_INTENT);
+                                intent.putExtra(MediaStore.EXTRA_OUTPUT,Uri.fromFile(imagen));
+                                startActivityForResult(intent,CAMERA_INTENT);
                                 break;
 
-                            case "Galería":
+                            case GALERIA:
                                 Log.i("perfil", "Selecciona galeria");
                                 Intent intent1 = new Intent(Intent.ACTION_PICK, MediaStore.Images.Media.EXTERNAL_CONTENT_URI);
                                 intent1.setType("image/*");
-                                startActivityForResult(intent1.createChooser(intent1, "Selecciona foto"), GALLERY_INTENT);
+                                startActivityForResult(intent1.createChooser(intent1,getString(R.string.selec_app)),GALLERY_INTENT);
 
                                 break;
 
-                            case "Cancelar":
+                            case CANCELAR:
                                 dialogInterface.dismiss();
                                 break;
                         }
@@ -121,12 +145,22 @@ public class Perfil extends AppCompatActivity {
                 camara = true;
                 galeria = false;
                 cambios = true;
-                Bundle bundle = data.getExtras();
-                final Bitmap bmp = (Bitmap) bundle.get("data");
-                Log.i("perfil", "requestCode==CAMERA_INTENT");
-                imagenUsuario.setImageBitmap(bmp);
-                stream = new ByteArrayOutputStream();
-                bmp.compress(Bitmap.CompressFormat.PNG, 100, stream);
+                MediaScannerConnection.scanFile(this, new String[]{miPath}, null,
+                        new MediaScannerConnection.OnScanCompletedListener() {
+                            @Override
+                            public void onScanCompleted(String s, Uri uri) {
+                                Log.i("Ruta de almacenamiento","Path -> " + miPath);
+                            }
+                        });
+                //Bitmap bitmap = BitmapFactory.decodeFile(miPath);
+                //avatarIcono.setImageBitmap(bitmap);
+                Glide.with(this)
+                        .load(miPath)
+                        .centerCrop()
+                        .fitCenter()
+                        .into(imagenUsuario);
+               // stream = new ByteArrayOutputStream();
+               // bitmap.compress(Bitmap.CompressFormat.PNG, 100, stream);
 
 
             } else if (requestCode == GALLERY_INTENT) {
@@ -148,7 +182,9 @@ public class Perfil extends AppCompatActivity {
 
     public void modificarPerfil(View view) {
         botonModificar.setClickable(false);
-
+        Animation animation = AnimationUtils.loadAnimation(getApplicationContext(), R.anim.scale);
+        animation.setDuration(1000);
+        botonModificar.setAnimation(animation);
         filePath = storageReference.child(FirebaseReferences.FOTO_PERFIL_USUARIO + "/" + editTextEmailUsuario.getText().toString() + "_" + getFecha());
 
         FirebaseDatabase database = FirebaseDatabase.getInstance();
@@ -158,7 +194,13 @@ public class Perfil extends AppCompatActivity {
 
         if (camara) {
             Log.i("perfil ", "entra camara");
-            byte[] data = stream.toByteArray();
+           // byte[] data = stream.toByteArray();
+            imagenUsuario.setDrawingCacheEnabled(true);
+            imagenUsuario.buildDrawingCache();
+            Bitmap bitmap = imagenUsuario.getDrawingCache();
+            ByteArrayOutputStream byteArrayOutputStream = new ByteArrayOutputStream();
+            bitmap.compress(Bitmap.CompressFormat.PNG, 100, byteArrayOutputStream);
+            byte[] data = byteArrayOutputStream.toByteArray();
             UploadTask uploadTask = filePath.putBytes(data);
             uploadTask.addOnSuccessListener(new OnSuccessListener<UploadTask.TaskSnapshot>() {
                 @Override
@@ -187,7 +229,7 @@ public class Perfil extends AppCompatActivity {
                 }
             });
         } else {
-            Toast.makeText(this, "Perfil modificado correctamente", Toast.LENGTH_SHORT).show();
+            Toast.makeText(this, R.string.perfil_mod, Toast.LENGTH_SHORT).show();
             cambios = false;
             botonModificar.setClickable(true);
             //finish();
@@ -197,21 +239,29 @@ public class Perfil extends AppCompatActivity {
     private void modificarFotoPerfil() {
 
         StorageReference photoRef = FirebaseStorage.getInstance().getReferenceFromUrl(user.getAvatar());
-        photoRef.delete().addOnSuccessListener(new OnSuccessListener<Void>() {
-            @Override
-            public void onSuccess(Void aVoid) {
+        Log.i("NoBorrar"," " + photoRef.getName());
+        Log.i("NoBorrar"," " + photoRef.toString());
+        if(!photoRef.getName().equals("series_ic.png")){
+            photoRef.delete().addOnSuccessListener(new OnSuccessListener<Void>() {
+                @Override
+                public void onSuccess(Void aVoid) {
 
-                Log.i("Perfil", "foto borrada correctamente");
-                FirebaseDatabase d = FirebaseDatabase.getInstance();
-                DatabaseReference refef = d.getReference().child(FirebaseReferences.USUARIOS_REFERENCE).child(ComunicarClaveUsuarioActual.getClave())
-                        .child(FirebaseReferences.AVATAR);
-                refef.setValue(enlaceFotoFirebase.toString());
-                Toast.makeText(Perfil.this, "Perfil modificado correctamente", Toast.LENGTH_SHORT).show();
-                cambios = false;
-                botonModificar.setClickable(true);
-                //finish();
-            }
-        });
+                    Log.i("Perfil", "foto borrada correctamente");
+
+                }
+            });
+        }
+        FirebaseDatabase d = FirebaseDatabase.getInstance();
+        DatabaseReference refef = d.getReference().child(FirebaseReferences.USUARIOS_REFERENCE).child(ComunicarClaveUsuarioActual.getClave())
+                .child(FirebaseReferences.AVATAR);
+        refef.setValue(enlaceFotoFirebase.toString());
+        Toast.makeText(Perfil.this,  R.string.perfil_mod, Toast.LENGTH_SHORT).show();
+        cambios = false;
+        botonModificar.setClickable(true);
+        //finish();
+
+
+
 
 
     }
@@ -243,16 +293,16 @@ public class Perfil extends AppCompatActivity {
                         }
                         if (cambios) {
                             AlertDialog.Builder dialogo1 = new AlertDialog.Builder(Perfil.this);
-                            dialogo1.setTitle("Importante");
-                            dialogo1.setMessage("¿Desea guardar los cambios?");
+                            dialogo1.setTitle(R.string.importante);
+                            dialogo1.setMessage(R.string.guardarcambios);
                             dialogo1.setCancelable(false);
-                            dialogo1.setPositiveButton("Confirmar", new DialogInterface.OnClickListener() {
+                            dialogo1.setPositiveButton(R.string.confirmar, new DialogInterface.OnClickListener() {
                                 public void onClick(DialogInterface dialogo1, int id) {
                                     modificarPerfil(null);
                                     finish();
                                 }
                             });
-                            dialogo1.setNegativeButton("Cancelar", new DialogInterface.OnClickListener() {
+                            dialogo1.setNegativeButton(R.string.cancelar, new DialogInterface.OnClickListener() {
                                 public void onClick(DialogInterface dialogo1, int id) {
                                     finish();
                                 }
@@ -294,16 +344,16 @@ public class Perfil extends AppCompatActivity {
                         }
                         if (cambios) {
                             AlertDialog.Builder dialogo1 = new AlertDialog.Builder(Perfil.this);
-                            dialogo1.setTitle("Importante");
-                            dialogo1.setMessage("¿Desea guardar los cambios?");
+                            dialogo1.setTitle(R.string.importante);
+                            dialogo1.setMessage(R.string.guardarcambios);
                             dialogo1.setCancelable(false);
-                            dialogo1.setPositiveButton("Confirmar", new DialogInterface.OnClickListener() {
+                            dialogo1.setPositiveButton(R.string.confirmar, new DialogInterface.OnClickListener() {
                                 public void onClick(DialogInterface dialogo1, int id) {
                                     modificarPerfil(null);
                                     finish();
                                 }
                             });
-                            dialogo1.setNegativeButton("Cancelar", new DialogInterface.OnClickListener() {
+                            dialogo1.setNegativeButton(R.string.cancelar, new DialogInterface.OnClickListener() {
                                 public void onClick(DialogInterface dialogo1, int id) {
                                     finish();
                                 }
